requirements.txt
boto3>=1.26.0
requests>=2.28.0
openai==0.28.1
pinecone>=3.0
python-dateutil>=2.8.2
openai[datalib]